Move formatting methods to utils
